refactored receiver (#4)
